% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_row_header}
\alias{add_row_header}
\title{Add row header helper function}
\usage{
add_row_header(kbl_object, table_to_print, search_term, extra = NULL)
}
\arguments{
\item{kbl_object}{Kable object}

\item{table_to_print}{Table used in Kable object}

\item{search_term}{Label to search for in table and row header text}

\item{extra}{Extra text to print in row header}
}
\value{
If search_term present in term, modified Kable object,
otherwise: original object is returned
}
\description{
A function that adds row headers to kable object
}
\keyword{internal}
