% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casewise_log_likelihood.R
\name{casewise_log_likelihood}
\alias{casewise_log_likelihood}
\title{Get casewise log-likelihood when full data is available}
\usage{
casewise_log_likelihood(clean_results, include_residuals = FALSE)
}
\arguments{
\item{clean_results}{A model fitted with minorbsem}

\item{include_residuals}{(LOGICAL) TRUE: Include minor factor
residual covariances in model-implied covariance matrix;
FALSE: Exclude them. If TRUE, different
models fit to the data will hardly be distinguishable.}
}
\value{
matrix (posterior iterations BY sample size)
containing log-likelihood
}
\description{
Get casewise log-likelihood for complete data,
useful for WAIC, LOOIC, ...
}
\examples{
# Comparing two models using LOOCV
fit_1 <- minorbsem("F1 =~ x1 + x2 + x3
                    F2 =~ x4 + x5 + x6
                    F3 =~ x7 + x8 + x9", HS)
# Compute case wise log-likelihood, exclude minor factor residuals
ll_mat_1 <- casewise_log_likelihood(fit_1, include_residuals = FALSE)
chain_id <- posterior::as_draws_df(fit_1$stan_fit)$.chain
loo_1 <- loo::loo(
  ll_mat_1,
  r_eff = loo::relative_eff(ll_mat_1, chain_id = chain_id)
)
print(loo_1)
fit_2 <- minorbsem("F1 =~ x1 + x2 + x3 + x9
                    F2 =~ x4 + x5 + x6
                    F3 =~ x7 + x8 + x9", HS)
# Compute case wise log-likelihood, exclude minor factor residuals
ll_mat_2 <- casewise_log_likelihood(fit_2, include_residuals = FALSE)
loo_2 <- loo::loo(
  ll_mat_2,
  r_eff = loo::relative_eff(ll_mat_2, chain_id = chain_id)
)
print(loo_2)
# Compare both models
print(loo::loo_compare(loo_1, loo_2), simplify = FALSE)
}
