% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_results.R
\name{ci_results}
\alias{ci_results}
\title{Conditional indepepence estimates for path analysis}
\usage{
ci_results(object, interval = 0.9, summarize = TRUE)
}
\arguments{
\item{object}{(mbsem) An object of \code{\link{mbsem-class}}
returned by \code{\link{minorbsem}}.}

\item{interval}{Confidence interval to select}

\item{summarize}{(LOGICAL)
If TRUE (default): Return posterior summary as data.frame;
If FALSE: Return posterior draws as data.frame.}
}
\value{
Table of parameters related to conditional independence.
}
\description{
Obtain conditional indepepence estimates for path analysis.
Estimates are standardized.
}
\examples{
\dontrun{
fit <- minorbpa("x3 ~ x1 + x2\n x4 ~ x3 + x1", HS)
ci_table <- ci_results(fit)
}
}
