% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_up_stan_fit}
\alias{clean_up_stan_fit}
\title{Clean up Stan fit helper function}
\usage{
clean_up_stan_fit(stan_fit, data_list)
}
\arguments{
\item{stan_fit}{Stan fit}

\item{data_list}{Data list object passed to Stan}
}
\value{
A list of major structural parameters,
minor factors standardized residual covariance matrix, and original Stan fit
}
\description{
A function that cleans up model returned by Stan
}
\keyword{internal}
