% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{converter_helper}
\alias{converter_helper}
\title{converter function}
\usage{
converter_helper(search_term, str_list)
}
\arguments{
\item{search_term}{Integer or string or NULL}

\item{str_list}{Character vector of accepted values}
}
\value{
If search_term is integer, returns string and vice-versa
}
\description{
A function that swaps type from string to integer
and vice-versa
}
\keyword{internal}
