% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list.R
\name{create_data_list}
\alias{create_data_list}
\title{Stan data helper function}
\usage{
create_data_list(
  lavaan_object = NULL,
  method = "normal",
  simple_struc = TRUE,
  priors = NULL,
  model = NULL,
  orthogonal = NULL
)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", "WB", "WB-cond", "WW", or "none". See details below.}

\item{simple_struc}{(LOGICAL) Only relevant for CFAs.
If TRUE: assume simple structure;
If FALSE: estimate all cross-loadings using generalized
double Pareto priors.}

\item{priors}{An object of \code{\link{mbsempriors-class}}.
See \code{\link{new_mbsempriors}} for more information.}

\item{model}{A description of the user-specified model, lavaan syntax.}

\item{orthogonal}{(logical) constrain factors orthogonal, must be TRUE to fit
bifactor models.}
}
\value{
Data list object used in fitting Stan model
}
\description{
A function that creates data list object passed to Stan
}
\keyword{internal}
