% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_major_params}
\alias{create_major_params}
\title{Create major parameters helper function}
\usage{
create_major_params(stan_fit, data_list, interval = 0.9)
}
\arguments{
\item{stan_fit}{Fitted Stan object}

\item{data_list}{Data list object passed to Stan}

\item{interval}{Confidence interval to select}
}
\value{
Summary of posterior draws
}
\description{
A function that creates the table of major parameters
}
\keyword{internal}
