% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_mi_vcov_ll}
\alias{create_mi_vcov_ll}
\title{Create model implied cov matrix or log-likelihood helper function}
\usage{
create_mi_vcov_ll(mat, data_list, include_residuals, return_ll = FALSE)
}
\arguments{
\item{mat}{Matrix of posterior samples}

\item{data_list}{Data list object passed to Stan}

\item{include_residuals}{(LOGICAL)
TRUE: Include minor factor residual covariances
in model-implied covariance matrix; FALSE: Exclude them. If TRUE, different
models fit to the data will hardly be distinguishable.}

\item{return_ll}{(LOGICAL)
TRUE: Return matrix of log-likelihood
FALSE: Return matrix of covariances}
}
\value{
Returns a matrix, which one depends on
\code{return_ll} argument.
}
\description{
Create model implied cov matrix or log-likelihood helper function
}
\keyword{internal}
