% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_model_implied_vcov}
\alias{create_model_implied_vcov}
\title{Create model implied covariance matrix fit helper function}
\usage{
create_model_implied_vcov(mat, data_list, include_residuals)
}
\arguments{
\item{mat}{Matrix of posterior samples}

\item{data_list}{Data list object passed to Stan}

\item{include_residuals}{(LOGICAL)
TRUE: Include minor factor residual covariances
in model-implied covariance matrix; FALSE: Exclude them. If TRUE, different
models fit to the data will hardly be distinguishable.}
}
\value{
A matrix that is ready for plotting
}
\description{
Create model implied covariance matrix fit helper function
}
\keyword{internal}
