% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_single_cfa_vcov_row}
\alias{create_single_cfa_vcov_row}
\title{Create model implied covariance matrix from CFA, helper function}
\usage{
create_single_cfa_vcov_row(
  params,
  data_list,
  include_residuals,
  all_lo,
  all_ev,
  all_ph
)
}
\arguments{
\item{params}{vector of posterior samples from a single iteration}

\item{data_list}{Data list object passed to Stan}

\item{include_residuals}{(LOGICAL)
TRUE: Include minor factor residual covariances
in model-implied covariance matrix;
FALSE: Exclude them. If TRUE, different
models fit to the data will hardly be distinguishable.}

\item{all_lo}{loading indexes}

\item{all_ev}{error variance indexes}

\item{all_ph}{factor correlation indexes}
}
\value{
A single model-impled covariance matrix
}
\description{
Create model implied covariance matrix from CFA, helper function
}
\keyword{internal}
