% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fix_acov}
\alias{.fix_acov}
\title{A function to reorder asymptotic variance matrix}
\usage{
.fix_acov(acov_mat, old_names, new_names)
}
\arguments{
\item{acov_mat}{Major paramters table}

\item{old_names}{Old names from correlation matrix}

\item{new_names}{Old names from lavaan object}
}
\value{
Returns reordered acov_mat
}
\description{
A function to reorder asymptotic variance matrix
}
\keyword{internal}
