% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.four_idx_solver}
\alias{.four_idx_solver}
\title{An internal solver for omega}
\usage{
.four_idx_solver(r_mat, idx_four)
}
\arguments{
\item{r_mat}{(matrix) Correlation matrix}

\item{idx_four}{(array) Contains four-way indices for computing omega.}
}
\value{
asymptotic covariances
}
\description{
An internal solver for omega
}
\keyword{internal}
