% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.g_map}
\alias{.g_map}
\title{Log matrix function using eigendecomposition}
\usage{
.g_map(r_vec)
}
\arguments{
\item{r_vec}{(vector) Strict lower half vector of correlation matrix}
}
\value{
Strict lower half vector of log(correlation matrix)
}
\description{
Log matrix function using eigendecomposition
}
\keyword{internal}
