% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_asy_cov}
\alias{.get_asy_cov}
\title{Asymptotic variance matrix of lower half vector of correlation matrix}
\usage{
.get_asy_cov(r_mat)
}
\arguments{
\item{r_mat}{(matrix) Correlation matrix}
}
\value{
Asymptotic variance matrix of strict lower half vector
of correlation matrix.
}
\description{
Asymptotic variance matrix of lower half vector of correlation matrix
}
\keyword{internal}
