% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.log_m}
\alias{.log_m}
\title{Log matrix function using eigendecomposition}
\usage{
.log_m(r_mat)
}
\arguments{
\item{r_mat}{(matrix) Correlation matrix}
}
\value{
Logarithm of the correlation matrix
}
\description{
Log matrix function using eigendecomposition
}
\keyword{internal}
