% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.omega_computer}
\alias{.omega_computer}
\title{Asymptotic variance matrix of lower half vector of correlation matrix}
\usage{
.omega_computer(r_mat, p_ast, ltri_idxs)
}
\arguments{
\item{r_mat}{(matrix) Correlation matrix}

\item{p_ast}{(positive integer) Dimension of strict lower half vector}

\item{ltri_idxs}{(array) Indices of strict lower half vector}
}
\value{
Asymptotic variance matrix of strict lower half vector
of correlation matrix.
}
\description{
Asymptotic variance matrix of lower half vector of correlation matrix
}
\keyword{internal}
