% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_in_missing_covs}
\alias{fill_in_missing_covs}
\title{Fill in missing values in sample covs function}
\usage{
fill_in_missing_covs(sample_cov)
}
\arguments{
\item{sample_cov}{(list of matrices) sample variance-covariance matrices.
The rownames and/or colnames must contain the observed variable names.
For now, assumes there are no missing elements in the covariance matrices.}
}
\value{
list of filled in sample covs
}
\description{
Fill in missing values in sample covs function
}
\keyword{internal}
