% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_plot_list}
\alias{get_param_plot_list}
\title{Create parameter list for plotting}
\usage{
get_param_plot_list(data_list)
}
\arguments{
\item{data_list}{Data list object passed to Stan}
}
\value{
Name of varying model parameters
}
\description{
Create parameter list for plotting
}
\keyword{internal}
