% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS}
\alias{HS}
\title{Holzinger-Swineford data}
\format{
\subsection{\code{HS}}{

A data frame with 301 observations and 15 variables:
\describe{
\item{id}{Identifier}
\item{sex}{Gender}
\item{ageyr}{Age, year part}
\item{agemo}{Age, month part}
\item{school}{School (Pasteur or Grant-White)}
\item{grade}{Grade}
\item{x1}{Visual perception}
\item{x2}{Cubes}
\item{x3}{Lozenges}
\item{x4}{Paragraph comprehension}
\item{x5}{Sentence completion}
\item{x6}{Word meaning}
\item{x7}{Speeded addition}
\item{x8}{Speeded counting of dots}
\item{x9}{Speeded discrimination straight and curved capitals}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/lavaan/index.html}
}
\usage{
HS
}
\description{
Reduced form data retrieved from lavaan
}
\keyword{datasets}
