% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{include_residuals}
\alias{include_residuals}
\title{Include minor factor residuals in model implied covariance matrix, helper function}
\usage{
include_residuals(Omega, params, data_list)
}
\arguments{
\item{Omega}{model implied covariance matrix to be updated}

\item{params}{vector of posterior samples from a single iteration}

\item{data_list}{Data list object passed to Stan}
}
\value{
A single model-impled covariance matrix
}
\description{
Include minor factor residuals in model implied covariance matrix, helper function
}
\keyword{internal}
