% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{issp89}
\alias{issp89}
\title{Another dataset from the metaSEM package.}
\format{
\subsection{\code{issp89}}{

A list of data with the following structure:
\describe{
\item{data}{A list of 11 studies of covariance matrices}
\item{n}{A vector of sample sizes}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/metaSEM/index.html}
}
\usage{
issp89
}
\description{
The data set includes 11 studies on 9 items measuring work-related attitudes
\insertCite{noauthor_international_1992}{minorbsem}.
Data and documentation copied over from the metaSEM package
\insertCite{Cheung-metaSEM}{minorbsem}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
