% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mbsem_post_sum}
\alias{mbsem_post_sum}
\title{Posterior summary helper function}
\usage{
mbsem_post_sum(stan_fit, variable, interval = 0.9, major = FALSE)
}
\arguments{
\item{stan_fit}{Fitted Stan object}

\item{variable}{Variable(s) to search for in Stan}

\item{interval}{Confidence interval to select}

\item{major}{If TRUE, add some preamble for printing the major
parameters table.}
}
\value{
Summary of posterior draws
}
\description{
A function that slightly modifies the default
summary function in posterior package
}
\keyword{internal}
