% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_defintions.R
\docType{class}
\name{mbsempriors-class}
\alias{mbsempriors-class}
\title{A class for setting up priors.}
\description{
A class for setting up priors.
}
\section{Slots}{

\describe{
\item{\code{lkj_shape}}{(positive real) The shape parameter of the LKJ-prior on the
interfactor correlation matrix in confirmatory factor models.}

\item{\code{sl_par}}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
deviation of loadings.}

\item{\code{fs_par}}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the standard deviation
of factors in SEMs.}

\item{\code{rs_par}}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the residual standard deviations.}

\item{\code{rc_par}}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
prior on the residual error correlations.}

\item{\code{sc_par}}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
deviations of coefficients; SD(coefs) vary by outcome.}

\item{\code{fc_par}}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
prior on the inter-factor correlations in latent regression models.}
}}

