% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meta_covs_check}
\alias{meta_covs_check}
\title{Check meta-user sample covariances input function}
\usage{
meta_covs_check(sample_cov)
}
\arguments{
\item{sample_cov}{(list of matrices) sample variance-covariance matrices.
The rownames and/or colnames must contain the observed variable names.}
}
\description{
A function that checks meta-user sample cov for adequacy
and fails on inadequate input.
}
\keyword{internal}
