% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meta_nobs_check}
\alias{meta_nobs_check}
\title{Check meta-user number input function}
\usage{
meta_nobs_check(sample_nobs)
}
\arguments{
\item{sample_nobs}{(vector of positive integer) Number of observations
for each study.}
}
\description{
A function that checks meta-user sample nobs for adequacy
and fails on inadequate input.
}
\keyword{internal}
