% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{method_hash}
\alias{method_hash}
\title{Method hash function}
\usage{
method_hash(search_term = NULL)
}
\arguments{
\item{search_term}{Integer or string or NULL}
}
\value{
If search_term is integer, returns string and vice-versa
}
\description{
A function that swaps method from string to integer
and vice-versa
}
\keyword{internal}
