% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minorbsem.R
\name{minorbsem}
\alias{minorbsem}
\title{Fit Bayesian SEMs with minor factors assumed}
\usage{
minorbsem(
  model = NULL,
  data = NULL,
  sample_cov = NULL,
  sample_nobs = NULL,
  orthogonal = FALSE,
  seed = 12345,
  warmup = 500,
  sampling = 500,
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 3,
  ncores = parallel::detectCores() - 2,
  lkj_shape = 2,
  sl_par = 1,
  rs_par = 2.5,
  rc_par = 2,
  sc_par = 1
)
}
\arguments{
\item{model}{A description of the user-specified model, lavaan syntax.}

\item{data}{An optional data frame containing the observed variables used in
the model.}

\item{sample_cov}{(matrix) sample variance-covariance matrix.
The rownames and/or colnames must contain the observed variable names.}

\item{sample_nobs}{(positive integer) Number of observations if the full
data frame
is missing and only sample covariance matrix is given.}

\item{orthogonal}{(logical) constrain factors orthogonal, must be TRUE to fit
bifactor models.}

\item{seed}{(positive integer) seed, set to obtain replicable results.}

\item{warmup}{(positive integer) The number of warmup iterations to run per
chain.}

\item{sampling}{(positive integer) The number of post-warmup iterations to
run per chain, retained for inference.}

\item{adapt_delta}{(real in (0, 1)) Increase to resolve divergent
transitions.}

\item{max_treedepth}{(positive integer) Increase to resolve problems with
maximum tree depth.}

\item{chains}{(positive integer) The number of Markov chains to run.}

\item{ncores}{(positive integer) The number of chains to run in parallel.}

\item{lkj_shape}{(positive real) The shape parameter of the LKJ-prior on the
interfactor correlation matrix.}

\item{sl_par}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
deviation of loadings.}

\item{rs_par}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the residual standard deviations.}

\item{rc_par}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
prior on the residual error correlations.}

\item{sc_par}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
deviations of coefficients; SD(coefs) vary by outcome.}
}
\value{
A list
}
\description{
Fit Bayesian SEMs with minor factors assumed
}
\examples{
minorbsem("# latent variable definitions
           F1 =~ x1 + x2 + x3
           F2 =~ x4 + x5 + x6
           F3 =~ x7 + x8 + x9", HS)
minorbsem("# latent variable definitions
           ind60 =~ x1 + x2 + x3
           dem60 =~ y1 + y2 + y3 + y4
           dem65 =~ y5 + y6 + y7 + y8
           # latent regressions
           dem60 ~ ind60
           dem65 ~ ind60 + dem60", PD)
}
