% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minorbsem_version}
\alias{minorbsem_version}
\title{Get package version function}
\usage{
minorbsem_version()
}
\value{
Package version as string
}
\description{
Get package version function
}
\keyword{internal}
