% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_major_params}
\alias{modify_major_params}
\title{Modify major parameters table helper function}
\usage{
modify_major_params(
  major_parameters,
  idxs,
  group = "",
  op = "",
  from = "",
  to = ""
)
}
\arguments{
\item{major_parameters}{Major paramters table}

\item{idxs}{Relevant rows indexes}

\item{group}{Parameter group}

\item{op}{lavaan style operator}

\item{from}{Variable from}

\item{to}{Variable to}
}
\value{
Updated major parameters table
}
\description{
A function that adds user friendly descriptions to the
major parameters table
}
\keyword{internal}
