% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Norton13}
\alias{Norton13}
\title{Studies on the Hospital Anxiety and Depression Scale.}
\format{
\subsection{\code{Norton13}}{

A list of data with the following structure:
\describe{
\item{data}{
A list of 28 studies of correlation matrices.
The variables are 14 items (x1 to x14) measuring HADS}
\item{n}{A vector of sample sizes}
\item{population}{A vector of the population of the data}
\item{group}{
A vector of classification into patients vs. non-patients
based on population}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/metaSEM/index.html}
}
\usage{
Norton13
}
\description{
The data set includes 28 studies on 14 items measuring the Hospital Anxiety
and Depression Scale (HADS) Reported by
\insertCite{norton_hospital_2013;textual}{minorbsem}, with data and
documentation copied over from the metaSEM package
\insertCite{Cheung-metaSEM}{minorbsem}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
