% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_hist.R
\name{parameter_hist}
\alias{parameter_hist}
\title{Histogram of parameter posterior distribution}
\usage{
parameter_hist(object, subset = NULL, ...)
}
\arguments{
\item{object}{(mbsem) An object of \code{\link{mbsem-class}}
returned by \code{\link{minorbsem}}.}

\item{subset}{(character) Subset of parameters:
NULL (Default) showing all estimated parameters;
Any other response will be used as regular expressions to subset
the parameters. It can be loading names or types of parameters.}

\item{...}{additional arguments to relevant bayesplot function}
}
\value{
bayesplot object
}
\description{
Produce histograms of parameter posterior distribution,
option to limit plot by type of parameter.
}
\examples{
\dontrun{
fit <- minorbsem("F1 =~ x1 + x2 + x3
                  F2 =~ x4 + x5 + x6
                  F3 =~ x7 + x8 + x9", HS)
parameter_hist(fit)
}
}
