% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PD}
\alias{PD}
\title{Industrialization And Political Democracy Dataset}
\format{
\subsection{\code{PD}}{

A data frame with 75 observations and 11 variables:
\describe{
\item{y1}{Expert ratings of the freedom of the press in 1960}
\item{y2}{The freedom of political opposition in 1960}
\item{y3}{The fairness of elections in 1960}
\item{y4}{The effectiveness of the elected legislature in 1960}
\item{y5}{Expert ratings of the freedom of the press in 1965}
\item{y6}{The freedom of political opposition in 1965}
\item{y7}{The fairness of elections in 1965}
\item{y8}{The effectiveness of the elected legislature in 1965}
\item{x1}{The gross national product (GNP) per capita in 1960}
\item{x2}{The inanimate energy consumption per capita in 1960}
\item{x3}{The percentage of the labor force in industry in 1960}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/lavaan/index.html}
}
\usage{
PD
}
\description{
Dataset in Bollen (1989), also retrieved from lavaan
}
\keyword{datasets}
