% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_missing_data_list}
\alias{prepare_missing_data_list}
\title{Create missing data matrices function}
\usage{
prepare_missing_data_list(data_list = NULL)
}
\value{
list of missing data matrices
}
\description{
Create missing data matrices function
}
\keyword{internal}
