% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_summary.R
\name{pretty_print_summary}
\alias{pretty_print_summary}
\title{Pretty print model results}
\usage{
pretty_print_summary(
  object,
  interval = 0.9,
  digits = 3,
  simple = TRUE,
  save_html = NULL
)
}
\arguments{
\item{object}{(mbsem) An object of \code{\link{mbsem-class}}
returned by \code{\link{minorbsem}}.}

\item{interval}{(real in (0, 1)) Credible interval to return.}

\item{digits}{(positive integer) Number of decimal places to print in table}

\item{simple}{(Logical) TRUE to produce table with less information
about parameters;
FALSE: produces table with more information}

\item{save_html}{(string) Optional file name to save table as HTML}
}
\description{
Nice printing of model results,
optionally produces HTML document
}
\examples{
\dontrun{
fit <- minorbsem("F1 =~ x1 + x2 + x3
                  F2 =~ x4 + x5 + x6
                  F3 =~ x7 + x8 + x9", HS)
pretty_print_summary(fit)
pretty_print_summary(fit, simple = FALSE)
}
}
