% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_method_selection}
\alias{random_method_selection}
\title{Select random method and random case function}
\usage{
random_method_selection()
}
\value{
randomly selected method and case
}
\description{
Select random method and random case function
}
\keyword{internal}
