% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_post_df_columns}
\alias{rename_post_df_columns}
\title{Rename columns of posterior data.frame prior by parameter type}
\usage{
rename_post_df_columns(
  df,
  labels_1,
  labels_2,
  begin_name,
  operation,
  search_replace,
  search_term_1,
  search_term_2
)
}
\arguments{
\item{df}{Posterior data.frame}

\item{labels_1}{Labels for first variable}

\item{labels_2}{Labels for second variable}

\item{begin_name}{String to begin columns names with}

\item{operation}{Operation, one of =~, ~, ~~}

\item{search_replace}{TRUE/FALSE for type of column naming}

\item{search_term_1}{Terms to use in replacement for first variable}

\item{search_term_2}{Terms to use in replacement for second variable}
}
\value{
A renamed posterior data.frame
}
\description{
Rename columns of posterior data.frame prior by parameter type
}
\keyword{internal}
