% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{target_fitter}
\alias{target_fitter}
\title{target fitter function}
\usage{
target_fitter(
  data_list,
  seed,
  warmup,
  sampling,
  refresh,
  adapt_delta,
  max_treedepth,
  chains,
  ncores,
  show_messages,
  pa = FALSE
)
}
\arguments{
\item{data_list}{Data list object passed to Stan}

\item{seed}{(positive integer) seed, set to obtain replicable results.}

\item{warmup}{(positive integer) The number of warmup iterations to run per
chain.}

\item{sampling}{(positive integer) The number of post-warmup iterations to
run per chain, retained for inference.}

\item{refresh}{(positive integer) How often to print the status
of the sampler.}

\item{adapt_delta}{(real in (0, 1)) Increase to resolve divergent
transitions.}

\item{max_treedepth}{(positive integer) Increase to resolve problems with
maximum tree depth.}

\item{chains}{(positive integer) The number of Markov chains to run.}

\item{ncores}{(positive integer) The number of chains to run in parallel.}

\item{show_messages}{(Logical) If TRUE, show messages from Stan sampler,
if FALSE, hide messages.}

\item{pa}{(LOGICAL)
If TRUE: Path-analytic model;
If FALSE (default): Generic model.}
}
\value{
Fitted Stan model
}
\description{
A function that takes user input and fits the
Stan model.
}
\keyword{internal}
