% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cmdstan_loc.R
\name{update_cmdstan_loc}
\alias{update_cmdstan_loc}
\title{Update CmdStan location}
\usage{
update_cmdstan_loc(loc = "")
}
\arguments{
\item{loc}{(string) Folder CmdStan lives on system}
}
\description{
Update CmdStan location
}
\examples{
update_cmdstan_loc(loc = "~/cmdstan/")
}
