% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_sigma_mode}
\alias{use_sigma_mode}
\title{Use posterior mode of Sigma for WB, helper function}
\usage{
use_sigma_mode(omega_mat, params, data_list)
}
\arguments{
\item{omega_mat}{model implied covariance matrix to be updated}

\item{params}{vector of posterior samples from a single iteration}

\item{data_list}{Data list object passed to Stan}
}
\value{
A single model-impled covariance matrix
}
\description{
Use posterior mode of Sigma for WB, helper function
}
\keyword{internal}
