% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_input_check}
\alias{user_input_check}
\title{Check user input function}
\usage{
user_input_check(type, object_1 = NULL, object_2 = NULL, object_3 = NULL)
}
\arguments{
\item{type}{string for type of check}

\item{object_1}{Object to check}

\item{object_2}{Object to check}

\item{object_3}{Object to check}
}
\description{
A function that checks user input for adequacy
and fails on inadequate input.
}
\keyword{internal}
