% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_input_process}
\alias{user_input_process}
\title{User input processing function}
\usage{
user_input_process(
  model = NULL,
  data = NULL,
  sample_cov = NULL,
  sample_nobs = NULL,
  method = "normal",
  orthogonal = FALSE,
  simple_struc = TRUE,
  correlation = FALSE,
  centered = TRUE,
  priors = new_mbsempriors(),
  compute_ll = FALSE,
  acov_mat = NULL
)
}
\arguments{
\item{model}{A description of the user-specified model, lavaan syntax.}

\item{data}{An optional data frame containing the observed variables used in
the model.}

\item{sample_cov}{(matrix) sample variance-covariance matrix.
The rownames and/or colnames must contain the observed variable names.}

\item{sample_nobs}{(positive integer) Number of observations if the full
data frame
is missing and only sample covariance matrix is given.}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", "WB", "WB-cond", "WW", or "none". See details below.}

\item{orthogonal}{(logical) constrain factors orthogonal, must be TRUE to fit
bifactor models.}

\item{simple_struc}{(LOGICAL) Only relevant for CFAs.
If TRUE: assume simple structure;
If FALSE: estimate all cross-loadings using generalized
double Pareto priors.}

\item{correlation}{(LOGICAL)
If TRUE: perform correlation structure analysis based on logarithm
of a matrix transformation
\insertCite{archakov_new_2021}{minorbsem};
If FALSE (default): perform covariance structure analysis.}

\item{centered}{(LOGICAL)
Only relevant for WB-cond and WW methods when \code{correlation = TRUE}.
If TRUE (default): Use a centered parameterization;
If FALSE: Use a non-centered parameterization.}

\item{priors}{An object of \code{\link{mbsempriors-class}}.
See \code{\link{new_mbsempriors}} for more information.}

\item{compute_ll}{(Logical) If TRUE, compute log-likelihood,
if FALSE, do not. This may be useful for cross-validation. This argument is
ignored when:
(i) the full dataset is not provided;
(ii) the method is WB, use WB-cond instead.}

\item{acov_mat}{(Optional) Asymptotic variance matrix of
lower triangular half (column-order) of the correlation matrix
to be used for correlation structure analysis.
This parameter is useful if importing polychoric or meta-analytic
SEM pooled correlation matrix.}
}
\description{
A function that processes.
}
\keyword{internal}
