% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_param_type}
\alias{validate_param_type}
\title{Plotting param_type validation}
\usage{
validate_param_type(param_type)
}
\arguments{
\item{param_type}{param_type for plotting}
}
\description{
Plotting param_type validation
}
\keyword{internal}
