% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_stan_fit.R
\name{clean_up_stan_fit}
\alias{clean_up_stan_fit}
\title{Clean up Stan fit helper function}
\usage{
clean_up_stan_fit(stan_fit, data_list)
}
\arguments{
\item{stan_fit}{Stan fit}

\item{data_list}{Data list object passed to Stan}
}
\value{
An object of \code{\link{mbsem-class}}
}
\description{
A function that cleans up model returned by Stan
}
\keyword{internal}
