% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_residual_network.R
\name{compute_residual_network}
\alias{compute_residual_network}
\title{Compute residual network}
\usage{
compute_residual_network(object, ...)
}
\arguments{
\item{object}{(mbsem) An object of \code{\link{mbsem-class}}
returned by \code{\link{minorbsem}}.}
}
\value{
data.frame
}
\description{
Interpret the error correlations as a residual network model.
}
\examples{
\dontrun{
fit <- minorbsem("F1 =~ x1 + x2 + x3
                  F2 =~ x4 + x5 + x6
                  F3 =~ x7 + x8 + x9", HS)
res_net <- compute_residual_network(fit)
p_corr_df <- posterior::summarise_draws(res_net)
p_corr_mat <- matrix(p_corr_df$mean, fit@data_list$Ni)
p_corr_mat
qgraph::qgraph(p_corr_mat, threshold = .05)
}
}
