% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list.R
\name{create_data_list}
\alias{create_data_list}
\title{Stan data helper function}
\usage{
create_data_list(
  lavaan_object = NULL,
  method = "normal",
  simple_struc = TRUE,
  correlation = correlation,
  priors = NULL,
  compute_ll = FALSE,
  partab = NULL
)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", "WB", "WB-cond", "WW", or "none". See details below.}

\item{simple_struc}{(LOGICAL) Only relevant for CFAs.
If TRUE: assume simple structure;
If FALSE: estimate all cross-loadings using generalized
double Pareto priors.}

\item{correlation}{(LOGICAL)
If TRUE: perform correlation structure analysis based on logarithm
of a matrix transformation
\insertCite{archakov_new_2021}{minorbsem};
If FALSE (default): perform covariance structure analysis.}

\item{priors}{An object of \code{\link{mbsempriors-class}}.
See \code{\link{new_mbsempriors}} for more information.}

\item{compute_ll}{(Logical) If TRUE, compute log-likelihood,
if FALSE, do not. This may be useful for cross-validation. This argument is
ignored when:
(i) the full dataset is not provided;
(ii) the method is WB, use WB-cond instead.}

\item{partab}{lavaanify result of corresponding model}
}
\value{
Data list object used in fitting Stan model
}
\description{
A function that creates data list object passed to Stan
}
\keyword{internal}
