% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list.R
\name{create_data_list}
\alias{create_data_list}
\title{Stan data helper function}
\usage{
create_data_list(lavaan_object = NULL, method = "normal", priors = NULL)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{method}{(character) One of "normal", "lasso", "none"}

\item{priors}{An object of \code{\link{mbsempriors-class}}.
See \code{\link{new_mbsempriors}} for more information.}
}
\value{
Data list object used in fitting Stan model
}
\description{
A function that creates data list object passed to Stan
}
\keyword{internal}
