% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_data_list}
\alias{create_data_list}
\title{Stan data helper function}
\usage{
create_data_list(
  lavaan_object = NULL,
  lkj_shape = 2,
  sl_par = 1,
  rs_par = 2.5,
  rc_par = 2,
  sc_par = 1
)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{lkj_shape}{(positive real) The shape parameter of the LKJ-prior on the
interfactor correlation matrix.}

\item{sl_par}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
deviation of loadings.}

\item{rs_par}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the residual standard deviations.}

\item{rc_par}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
prior on the residual error correlations.}

\item{sc_par}{(positive real) The scale parameter of the
Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
deviations of coefficients; SD(coefs) vary by outcome.}
}
\value{
Data list object used in fitting Stan model
}
\description{
A function that creates data list object passed to Stan
}
\keyword{internal}
