% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minorbsem.R
\name{minorbsem}
\alias{minorbsem}
\title{Fit Bayesian SEMs with minor factors assumed}
\usage{
minorbsem(
  model = NULL,
  data = NULL,
  sample_cov = NULL,
  sample_nobs = NULL,
  method = "normal",
  orthogonal = FALSE,
  seed = 12345,
  warmup = 500,
  sampling = 500,
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 3,
  ncores = max(parallel::detectCores() - 2, 1),
  priors = new_mbsempriors(),
  show = TRUE
)
}
\arguments{
\item{model}{A description of the user-specified model, lavaan syntax.}

\item{data}{An optional data frame containing the observed variables used in
the model.}

\item{sample_cov}{(matrix) sample variance-covariance matrix.
The rownames and/or colnames must contain the observed variable names.}

\item{sample_nobs}{(positive integer) Number of observations if the full
data frame
is missing and only sample covariance matrix is given.}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP" or "none".
Select "normal" under belief that minor factor influences are
on average zero with continuous deviations away from zero.
Select "lasso" under belief that minor factor influences are
indeed zero with a small number of non-zero residual covariances.
Select "logistic" for similar belief as normal but more readily
accomodates extreme outliers.
Select "GDP" to mimic a global-local approach, i.e.
attempt to shrink near 0 residual covariances to 0
with minimal shrinking for larger residual covariances.
Select "none" if intending to ignore the influence of minor factors.}

\item{orthogonal}{(logical) constrain factors orthogonal, must be TRUE to fit
bifactor models.}

\item{seed}{(positive integer) seed, set to obtain replicable results.}

\item{warmup}{(positive integer) The number of warmup iterations to run per
chain.}

\item{sampling}{(positive integer) The number of post-warmup iterations to
run per chain, retained for inference.}

\item{adapt_delta}{(real in (0, 1)) Increase to resolve divergent
transitions.}

\item{max_treedepth}{(positive integer) Increase to resolve problems with
maximum tree depth.}

\item{chains}{(positive integer) The number of Markov chains to run.}

\item{ncores}{(positive integer) The number of chains to run in parallel.}

\item{priors}{An object of \code{\link{mbsempriors-class}}.
See \code{\link{new_mbsempriors}} for more information.}

\item{show}{(Logical) If TRUE, show table of results, if FALSE, do not
show table of results. As an example, use FALSE for simulation studies.}
}
\value{
An object of \code{\link{mbsem-class}}
}
\description{
The major function to fit models
}
\examples{
\dontrun{
minorbsem("# latent variable definitions
           F1 =~ x1 + x2 + x3
           F2 =~ x4 + x5 + x6
           F3 =~ x7 + x8 + x9", HS)
minorbsem("# latent variable definitions
           ind60 =~ x1 + x2 + x3
           dem60 =~ y1 + y2 + y3 + y4
           dem65 =~ y5 + y6 + y7 + y8
           # latent regressions
           dem60 ~ ind60
           dem65 ~ ind60 + dem60", PD)
}
}
