% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_hist.R
\name{parameter_hist}
\alias{parameter_hist}
\title{Histogram of parameter posterior distribution}
\usage{
parameter_hist(object, param_type = c("rm", "co", "lo", "fc", "fv"))
}
\arguments{
\item{object}{(mbsem) An object of class mbsem
returned by minorbsem.}

\item{param_type}{(string vector) Choose from a list of options:
"all" = all structural parameters;
"rm" = Root Mean square error of standardized residual covariances;
"lo" = LOadings;
"ev" = Error Variances;
"rc" = Residual Correlations;
"fc" = Factor Correlations;
"fv" = Factor Variances;
"co" = latent regression COefficients;
"re" = standardized REsidual covariances}
}
\value{
ggplot object
}
\description{
Produce histograms of parameter posterior distribution,
option to limit plot by type of parameter.
}
\examples{
fit <- minorbsem("F1 =~ x1 + x2 + x3
                  F2 =~ x4 + x5 + x6
                  F3 =~ x7 + x8 + x9", HS)
parameter_hist(fit)
parameter_hist(fit, param_type = "all")
parameter_hist(fit, param_type = c("rm", "lo", "fc"))
}
