% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Visualize standardized residual covariances}
\usage{
plot_residuals(object, method = "matrix")
}
\arguments{
\item{object}{(mbsem_object) An object of class mbsem_object
returned by minorbsem.}

\item{method}{(string) Either: "range" for lineranges (by default)
or "matrix" for a matrix with point estimates in the lower half.}
}
\value{
ggplot object
}
\description{
Visualize distribution of standardized residual covariances
}
\examples{
fit <- minorbsem("F1 =~ x1 + x2 + x3
                  F2 =~ x4 + x5 + x6
                  F3 =~ x7 + x8 + x9", HS)
plot_residuals(fit)
}
