% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stan_plot_data.R
\name{prepare_stan_plot_data}
\alias{prepare_stan_plot_data}
\title{Pre-plotting Stan fit helper function}
\usage{
prepare_stan_plot_data(object)
}
\arguments{
\item{object}{(mbsem_object) An object of class mbsem_object
returned by minorbsem.}
}
\value{
A data.frame that is ready for plotting
}
\description{
A function that cleans up model results to prepare for plotting
}
\keyword{internal}
