% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_stan_plot_data}
\alias{prepare_stan_plot_data}
\title{Pre-plotting Stan fit helper function}
\usage{
prepare_stan_plot_data(stan_fit, data_list)
}
\arguments{
\item{stan_fit}{Stan fit}

\item{data_list}{Data list object passed to Stan}
}
\value{
A data.frame that is ready for plotting
}
\description{
A function that cleans up model results to prepare for plotting
}
\keyword{internal}
